@@ -1,43 +1,37 @@
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin')
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const Dotenv = require('dotenv-webpack');
 
 module.exports = {
-    mode: 'development',
-    entry: path.join(__dirname, 'src/index.js'),
+    entry: './src/index.js',
     output: {
-        filename: 'main.js',
-        path: path.join(__dirname, 'dist'),
+        path: path.join(__dirname, '/dist'),
+        filename: 'bundle.js',
+    },
+    devServer: {
+        port: 8081,
+        historyApiFallback: true,
+        host: '0.0.0.0',
+        open: false,
+        allowedHosts: ['etl.dev'],
     },
     module: {
         rules: [
             {
-                test: /\.js$/,
-                exclude: /(node_modules)/,
-                use: {
-                    loader: 'babel-loader',
-                    options: {
-                        presets: [
-                            '@babel/preset-env',
-                            '@babel/preset-react' // async await
-                        ],
-                        plugins: [
-                            '@babel/plugin-transform-runtime'
-                        ]
-                    }
-                }
-            }
-        ]
+                test: /\.jsx?$/,
+                exclude: /node_modules/,
+                loader: 'babel-loader',
+            },
+            {
+                test: /\.css$/,
+                use: ['style-loader', 'css-loader'],
+            },
+        ],
     },
     plugins: [
         new HtmlWebpackPlugin({
-            inject: true,
-            template: path.join(__dirname, 'index.html'),
-            filename:'index.html'  
-        })
+            template: path.join(__dirname, '/src/index.html'),
+        }),
+        new Dotenv(),
     ],
-    devServer: {
-        port: 9000,
-        allowedHosts: 'all',
-        historyApiFallback: true
-    }
-}
+};
